Explorer node bootstrapping
BRS_FULL_BOOTSTRAP(default: true) option that if set to on/true will bootstrap the node with all known peers on the node.
BRS_BOOTSTRAP_NETWORK(default: [".signum.network"]) lets you set in list form the nodes you want to bootstrap with. example: [".signum.network"]

Both BRS_FULL_BOOTSTRAP and AUTO_BOOTSTRAP_PEERS need to be true for the process to happen, this was a design choice since the process takes a while and is a blocking function
The process will write the bootstrap peers found to the .env file when it's completed. It will also write BRS_FULL_BOOTSTRAP=off to avoid the process happening again.

Signed-off-by: gittrekt <76663968+gittrekt@users.noreply.github.com>